+++ src/langchain_mcp_tools/__init__.py
@@ -6,4 +6,5 @@ from .langchain_mcp_tools import (
   McpServerCommandBasedConfig,
   McpServerUrlBasedConfig,
   SingleMcpServerConfig,
+  McpInitializationError,
 )